Relocating test to decrease generated test case count
